Move Distribution and Link function in rolch.base
