+++ src/rolch/base/__init__.py
@@ -1,0 +1,4 @@
+from .distribution import Distribution
+from .link import LinkFunction
+
+__all__ = ["Distribution", "LinkFunction"]
