+++ src/rolch/distributions/studentt.py
@@ -2,7 +2,7 @@ import numpy as np
 import scipy.special as sp
 import scipy.stats as st
 
-from rolch.abc import Distribution
+from rolch.base import Distribution
 from rolch.link import IdentityLink, LogLink, LogShiftTwoLink
